+++ samples/kotlin-multiplatform-common/src/main/kotlin/org/jlleitschuh/gradle/ktlint/sample/kotlin/multiplatform/main.kt
@@ -1,4 +1,4 @@
-package org.jlleitschuh.gradle.ktlint.sample.kotlin.common
+package org.jlleitschuh.gradle.ktlint.sample.kotlin.multiplatform
 
 /**
  * The main function is multi-platform, it depends on platform specific code.
